+++ src/routes/web.php
@@ -4,6 +4,7 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', 'BackupController@index')->name('backup-ui.index');
 Route::post('/create', 'BackupController@create')->name('backup-ui.create');
+Route::get('/status', 'BackupController@status')->name('backup-ui.status');
 Route::get('/download/{disk}/{path}', 'BackupController@download')
     ->where('path', '.*')
     ->name('backup-ui.download');